passport jwt authentication added

+++ controllers/api/v1/posts_api.js
@@ -21,18 +21,19 @@ module.exports.destroy = async function (req, res) {
   try {
     let post = await Post.findById(req.params.id);
 
-    // if (post.user == req.user.id) {
-    post.remove();
+    if (post.user == req.user.id) {
+      post.remove();
 
-    await Comment.deleteMany({ post: req.params.id });
+      await Comment.deleteMany({ post: req.params.id });
 
-    return res.json(200, {
-      message: "post and associated message deleted succesfully",
-    });
-    // } else {
-    //   req.flash("error", "You cannot delete this post!");
-    //   return res.redirect("back");
-    // }
+      return res.json(200, {
+        message: "post and associated message deleted succesfully",
+      });
+    } else {
+      return res.json(401, {
+        message: "you cannot delete this post",
+      });
+    }
   } catch (err) {
     return res.json(500, {
       message: "Internal Server Error !!",
